Refactor, implement storage provider
