+++ ts/examples/memoryProviders/src/sqlite/storageDb.ts
@@ -5,6 +5,8 @@ import path from "path";
 import * as knowLib from "knowledge-processor";
 import { createDb, tablePath } from "./common.js";
 import { createTextIndex } from "./textTable.js";
+import { createObjectTable } from "./objectTable.js";
+import { ObjectFolder, ObjectFolderSettings } from "typeagent";
 
 export interface StorageDb extends knowLib.StorageProvider {
     readonly rootPath: string;
@@ -20,8 +22,16 @@ export async function createStorageDb(
     return {
         rootPath,
         name,
+        createObjectFolder: _createObjectFolder,
         createTextIndex: _createTextIndex,
     };
+
+    async function _createObjectFolder<T>(
+        folderPath: string,
+        settings?: ObjectFolderSettings,
+    ): Promise<ObjectFolder<T>> {
+        return createObjectTable<T>(db, folderPath, settings);
+    }
 
     async function _createTextIndex<TSourceId extends knowLib.ValueType>(
         settings: knowLib.TextIndexSettings,
