+++ ts/packages/knowledgeProcessor/src/index.ts
@@ -6,6 +6,7 @@ export * from "./textStore.js";
 export * from "./textChunker.js";
 export * from "./storageProvider.js";
 export * from "./keyValueIndex.js";
+export * from "./knowledgeStore.js";
 export * from "./knowledgeIndex.js";
 export * from "./temporal.js";
 export * from "./modelCache.js";
